move recursion test to test_distributed
