+++ dask/tests/test_distributed.py
@@ -744,3 +744,21 @@ async def test_non_recursive_df_reduce(c, s, a, b):
     )
 
     assert (await c.compute(result)).val == 170
+
+
+def test_set_index_no_resursion_error(client):
+    # see: https://github.com/dask/dask/issues/8955
+    try:
+        ddf = (
+            dask.datasets.timeseries(
+                start="2000-01-01",
+                end="2000-07-01",
+                freq="12h"
+            )
+            .reset_index()
+            .astype({"timestamp": str})
+        )
+        ddf = ddf.set_index("timestamp", sorted=True)
+        ddf.compute()
+    except RecursionError as re:
+        raise(re)
